Android: get EXIF orientation from content:// URI on r24+
Android has restricted access to the file:// URIs underlying a shared
content URI a while ago. I'm not sure when exactly, and my google-fu is
weak tonight.

On the other hand, they introduced ExifInterface(InputStream) in r24.
This can be used even when we do not have permissions / access to the
original file:// URI that the content:// URI resolves to.

So if we are on Android r24+, we use this new API to obtain the EXIF
image orientation and to fix portraits being sideways.